+++ src/impure/onig.rs
@@ -1,57 +1,26 @@
-#[cfg(feature = "oniguruma")]
 use std::io::{self, BufRead};
 
-#[cfg(feature = "oniguruma")]
 use onig::{self};
-#[cfg(feature = "oniguruma")]
 pub type Regex = onig::Regex;
-#[cfg(feature = "oniguruma")]
 pub type RegexOptions = onig::RegexOptions;
-#[cfg(feature = "oniguruma")]
 pub type Syntax = onig::Syntax;
 
-#[cfg(not(feature = "oniguruma"))]
-pub type Regex = i64;
-
-#[cfg(feature = "oniguruma")]
 use super::super::{errors, PipeIntercepter, DEFAULT_CAP, trim_eol, msg_error, error_exit};
 
-#[cfg(not(feature = "oniguruma"))]
-use super::super::{errors, PipeIntercepter};
-
-#[cfg(feature = "oniguruma")]
 pub fn new_regex() -> Regex {
     Regex::new("").unwrap()
 }
 
-#[cfg(not(feature = "oniguruma"))]
-pub fn new_regex() -> Regex {
-    1
-}
-
-#[cfg(feature = "oniguruma")]
 pub fn new_option_multiline_regex(s: &str) -> Regex {
     Regex::with_options(s, RegexOptions::REGEX_OPTION_MULTILINE, Syntax::default())
         .unwrap_or_else(|e| error_exit(&e.to_string()))
 }
 
-#[cfg(not(feature = "oniguruma"))]
-pub fn new_option_multiline_regex(_s: &str) -> Regex {
-    1
-}
-
-#[cfg(feature = "oniguruma")]
 pub fn new_option_none_regex(s: &str) -> Regex {
     Regex::with_options(s, RegexOptions::REGEX_OPTION_NONE, Syntax::default())
         .unwrap_or_else(|e| error_exit(&e.to_string()))
 }
 
-#[cfg(not(feature = "oniguruma"))]
-pub fn new_option_none_regex(_s: &str) -> Regex {
-    1
-}
-
-#[cfg(feature = "oniguruma")]
 /// Handles regex onig ( -g -G )
 pub fn regex_onig_proc(
     ch: &mut PipeIntercepter,
@@ -94,18 +63,6 @@ pub fn regex_onig_proc(
     Ok(())
 }
 
-#[cfg(not(feature = "oniguruma"))]
-/// Handles regex onig ( -g -G )
-pub fn regex_onig_proc(
-    _ch: &mut PipeIntercepter,
-    _line: &Vec<u8>,
-    _re: &Regex,
-    _invert: bool,
-) -> Result<(), errors::TokenSendError> {
-    Ok(())
-}
-
-#[cfg(feature = "oniguruma")]
 pub fn regex_onig_line_proc(
     ch: &mut PipeIntercepter,
     re: &Regex,
@@ -144,15 +101,5 @@ pub fn regex_onig_line_proc(
             Err(e) => msg_error(&e.to_string()),
         }
     }
-    Ok(())
-}
-
-#[cfg(not(feature = "oniguruma"))]
-pub fn regex_onig_line_proc(
-    _ch: &mut PipeIntercepter,
-    _re: &Regex,
-    _invert: bool,
-    _line_end: u8,
-) -> Result<(), errors::TokenSendError> {
     Ok(())
 }
